moved data-manager into modules
