+++ app/src/renderer/config/AppRoutes.config.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import TableViewIcon from '@mui/icons-material/TableView';
 import SettingsInputComponentIcon from '@mui/icons-material/SettingsInputComponent';
 
-import DataManager from '@renderer/components/data-manager';
+import DataManager from '@src/renderer/modules/data-manager';
 import { PrincipalComponentsAnalysis } from '@src/renderer/modules/pca';
 
 interface Route {
