+++ app/src/renderer/modules/data-manager/components/DecimalsHandler.tsx
