+++ app/src/renderer/modules/data-manager/components/ScalingHandler.tsx
