+++ app/src/renderer/modules/data-manager/data-frame-viewer/CheckBox.tsx
