+++ app/src/renderer/modules/data-manager/data-frame-viewer/styles.ts
