+++ app/src/renderer/store/data-manager/types.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/named
 import { Dispatch as ReduxDispatch } from 'redux';
 
-import { DataFrame, DecimalsPrecision } from '@renderer/components/data-manager/data-frame-viewer/types';
+import { DataFrame, DecimalsPrecision } from '@src/renderer/modules/data-manager/data-frame-viewer/types';
 
 export enum DataManagerActionTypes {
   Loading = 'LOADING',
